+++ analyzer/src/main/java/nl/basjes/parse/useragent/annotate/UserAgentAnnotationAnalyzer.java
@@ -19,7 +19,6 @@ package nl.basjes.parse.useragent.annotate;
 
 import nl.basjes.parse.useragent.UserAgent;
 import nl.basjes.parse.useragent.UserAgentAnalyzer;
-import nl.basjes.parse.useragent.UserAgentAnalyzer.UserAgentAnalyzerBuilder;
 import nl.basjes.parse.useragent.analyze.InvalidParserConfigurationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,12 +32,36 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static nl.basjes.parse.useragent.UserAgentAnalyzer.DEFAULT_PARSE_CACHE_SIZE;
+
 public class UserAgentAnnotationAnalyzer<T> {
     private UserAgentAnnotationMapper<T> mapper = null;
     private UserAgentAnalyzer userAgentAnalyzer = null;
+    private int cacheSize = DEFAULT_PARSE_CACHE_SIZE;
     private static final Logger LOG = LoggerFactory.getLogger(UserAgentAnnotationAnalyzer.class);
 
     private final Map<String, List<Method>> fieldSetters = new HashMap<>();
+
+    public void disableCaching() {
+        setCacheSize(0);
+    }
+
+    /**
+     * Sets the new size of the parsing cache.
+     * Note that this will also wipe the existing cache.
+     *
+     * @param newCacheSize The size of the new LRU cache. As size of 0 will disable caching.
+     */
+    public void setCacheSize(int newCacheSize) {
+        cacheSize = newCacheSize > 0 ? newCacheSize : 0;
+        if (userAgentAnalyzer != null) {
+            userAgentAnalyzer.setCacheSize(cacheSize);
+        }
+    }
+
+    public int getCacheSize() {
+        return cacheSize;
+    }
 
     public void initialize(UserAgentAnnotationMapper<T> theMapper) {
         mapper = theMapper;
@@ -102,10 +125,14 @@ public class UserAgentAnnotationAnalyzer<T> {
             throw new InvalidParserConfigurationException("You MUST specify at least 1 field to extract.");
         }
 
-        UserAgentAnalyzerBuilder<?, ?> builder = UserAgentAnalyzer.newBuilder();
-        builder.hideMatcherLoadStats();
-        builder.withFields(fieldSetters.keySet());
-        userAgentAnalyzer = builder.build();
+        userAgentAnalyzer = UserAgentAnalyzer
+            .newBuilder()
+            .hideMatcherLoadStats()
+            .withCache(cacheSize)
+            .withFields(fieldSetters.keySet())
+            .dropTests()
+            .immediateInitialization()
+            .build();
     }
 
     public T map(T record) {